Add class to build unpublishing email
Takes arguements from the UnpublishHandlerService and builds an email
for each subscriber that is subscribed to the taxon that has been
unpublished.

Also creates a copy of the unpublishing message to send to
govuk-email-courtesy-copies@digital.cabinet-office.gov.uk. This email
is owned by GDS and means we can see the emails that are being sent.